Remove deprecated ConsistentHash
